+++ tests/jest/aggregate-messages.test.js
@@ -1,4 +1,3 @@
-const fs = require('fs');
 const path = require('path');
 const aggregateMessages = require('../../scripts/aggregate-translations/aggregate-messages');
 
@@ -9,7 +8,6 @@ const translationDirectories = [
   path.resolve(__dirname, 'fixtures', 'ned_modules', 'fixtures1', 'translations'),
 ];
 const locales = ['en', 'en-US', 'es'];
-const fileSystem = fs;
 
 const expectedMessages = {
   'Terra.fixtures.test': 'Test...',
@@ -18,7 +16,7 @@ const expectedMessages = {
 
 describe('aggregates translations messages', () => {
   it('returns empty messages when no translations directories are provided', () => {
-    const messages = aggregateMessages([], locales, fileSystem);
+    const messages = aggregateMessages([], locales);
 
     expect(messages).toHaveProperty('en', {});
     expect(messages).toHaveProperty('en-US', {});
@@ -26,14 +24,14 @@ describe('aggregates translations messages', () => {
   });
 
   it('logs a warning message if a translation file is not found', () => {
-    aggregateMessages([__dirname], ['en'], fileSystem);
+    aggregateMessages([__dirname], ['en']);
 
     // eslint-disable-next-line no-console
     expect(console.warn).toBeCalled();
   });
 
   it('aggregates the messages', () => {
-    const messages = aggregateMessages(translationDirectories, locales, fileSystem);
+    const messages = aggregateMessages(translationDirectories, locales);
 
     expect(messages).toHaveProperty('en', expectedMessages);
     expect(messages).toHaveProperty('en-US', expectedMessages);